wip: reduce screenshot ui complexity
